# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface AnthropicConversationInput {
  title: string;
}

export interface AnthropicMessageInput {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export interface TutorSession {
  conversationId: number;
  topicId: number;
  topicTitle: string;
  level: string;
}

export interface TutorSessionWithMessages {
  conversationId: number;
  topicId: number;
  topicTitle: string;
  level: string;
  messages: AnthropicMessage[];
}

export type TutorSessionInputLevel = typeof TutorSessionInputLevel[keyof typeof TutorSessionInputLevel];


export const TutorSessionInputLevel = {
  new: 'new',
  experienced: 'experienced',
} as const;

export interface TutorSessionInput {
  topicId: number;
  level: TutorSessionInputLevel;
}

export type TutorMessageInputMessageType = typeof TutorMessageInputMessageType[keyof typeof TutorMessageInputMessageType];


export const TutorMessageInputMessageType = {
  normal: 'normal',
  hint: 'hint',
  stuck: 'stuck',
  synthesis: 'synthesis',
} as const;

export interface TutorMessageInput {
  content: string;
  messageType: TutorMessageInputMessageType;
}