# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutor/sessions:
    post:
      operationId: createTutorSession
      tags: [tutor]
      summary: Start a new Socratic tutor session for a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorSessionInput"
      responses:
        "201":
          description: Created tutor session with conversation ID and opening message SSE stream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorSession"
  /tutor/sessions/{conversationId}/message:
    post:
      operationId: sendTutorMessage
      tags: [tutor]
      summary: Send a learner message in a tutor session (SSE stream response)
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorMessageInput"
      responses:
        "200":
          description: SSE stream of tutor response chunks
          content:
            text/event-stream: {}
  /tutor/sessions/{conversationId}:
    get:
      operationId: getTutorSession
      tags: [tutor]
      summary: Get a tutor session with full message history
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorSessionWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
components:
  schemas:
...
    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
...
    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    TutorSession:
      type: object
      properties:
        conversationId:
          type: integer
        topicId:
          type: integer
        topicTitle:
          type: string
        level:
          type: string
      required:
        - conversationId
        - topicId
        - topicTitle
        - level
    TutorSessionWithMessages:
      type: object
      properties:
        conversationId:
          type: integer
        topicId:
          type: integer
        topicTitle:
          type: string
        level:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - conversationId
        - topicId
        - topicTitle
        - level
        - messages
    TutorSessionInput:
      type: object
      properties:
        topicId:
          type: integer
        level:
          type: string
          enum: [new, experienced]
      required:
        - topicId
        - level
    TutorMessageInput:
      type: object
      properties:
        content:
          type: string
        messageType:
          type: string
          enum: [normal, hint, stuck, synthesis]
      required:
        - content
        - messageType