# lib/api-zod/src/generated/types/tutorMessageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TutorMessageInputMessageType } from './tutorMessageInputMessageType';

export interface TutorMessageInput {
  content: string;
  messageType: TutorMessageInputMessageType;
}