# lib/api-zod/src/generated/types/tutorSession.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TutorSession {
  conversationId: number;
  topicId: number;
  topicTitle: string;
  level: string;
}