# lib/api-zod/src/generated/types/tutorSessionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TutorSessionInputLevel } from './tutorSessionInputLevel';

export interface TutorSessionInput {
  topicId: number;
  level: TutorSessionInputLevel;
}