# lib/api-zod/src/generated/types/tutorSessionWithMessages.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnthropicMessage } from './anthropicMessage';

export interface TutorSessionWithMessages {
  conversationId: number;
  topicId: number;
  topicTitle: string;
  level: string;
  messages: AnthropicMessage[];
}